api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
python-multipart
requests
pydantic-settings
email-validator

